refactor(parquet): clarify reader cost model flow
